Rename related and customize components
The components for handling related objects and the customize hook
have previously been named either something referring to either
related or customize. But the current implementation doesn't actually
have a clear line between the two concepts and makes the mmostly
inseparable, so they are moved to a common namespace with consistent
naming.
