+++ controller/common/customize/hook.go
@@ -1,4 +1,4 @@
-package common
+package customize
 
 import (
 	"fmt"
